minimmerflow: try to interleave memory accesses and computation in CUDA kernels
This can be achieved passing the containers that store device data
to the low level kernel functions. In this way it is not necessary
to gather all the data needed by low level functions before calling
the functions. Instead, each function will access the data only when
it actually needs. This reduces the amount of memory stalls because,
since data access and computation are interleaved, the GPU is able to
access the data needed later on while doing some computations.