constants: mark host's constants using the constexpr keyword
Variables marked constexpr can be directly used in device code:

"Let 'V' denote a namespace scope variable or a class static member
variable that has been marked constexpr and that does not have execution
space annotations (e.g., __device__, __constant__, __shared__). V is
considered to be a host code variable.

If V is of scalar type other than long double and the type is not
volatile-qualified, the value of V can be directly used in device
code. In addition, if V is of a non-scalar type then scalar elements
of V can be used inside a constexpr __device__ or __host__ __device__
function, if the call to the function is a constant expression.
Device source code cannot contain a reference to V or take the address
of V."