euler: fix evaluation of maximum eigenvalue
Evaluation of maximum eigenvalue through reduction can only be
used if all threads are active. The reason being that the reduction
operation synchronizes the threads and this is possible only if
all the threads are involved in the reduction.

Threads will be all active in the current iteration if all the
threads have an interface to process (i..e, the last thread of
the block is associated with a global id that is lower than the
number of interfaces being processed).